feat: update search bar to work with storefronts
Add a select next to the search bar that allows selecting whether to search products or storefronts. The select automatically switches depending on if you're on the products path or the storefronts path. The search bar also now remembers your search, making it much easier to edit your search terms.